requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apimatic-core==0.2.3
apimatic-core-interfaces==0.1.3
apimatic-requests-client-adapter==0.1.4
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
deprecation==2.1.0
dunamai==1.17.0
enum34==1.1.10
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
fuuid==0.1.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-translate==3.11.1
googleapis-common-protos==1.59.0
google-generativeai==0.2.1
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
jsonpointer==2.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.7
packaging==23.1
pandas==2.0.2
poetry-dynamic-versioning==0.13.1
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.31.0
rsa==4.9
scipy==1.10.1
six==1.16.0
squareup==27.0.0.20230517
tiktoken==0.4.0
tomlkit==0.11.8
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.4
yarl==1.9.2
